backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.17,<4.0
django-cors-headers>=4.9,<5.0
django-filter>=25.2,<26.0
Pillow>=12.1,<13.0
psycopg[binary]>=3.3,<4.0
gunicorn>=25.0,<26.0
dj-database-url>=3.1,<4.0
whitenoise>=6.12,<7.0
